Convert to using webpack
