+++ package.json
@@ -1,0 +1,24 @@
+{
+  "name": "blockcatcher",
+  "version": "1.0.0",
+  "description": "",
+  "private": true,
+  "scripts": {
+    "build": "webpack --progress --color",
+    "build-prod": "webpack --mode production --no-devtool",
+    "watch": "webpack --progress --color --watch",
+    "dev": "webpack serve --progress --color",
+    "lint": "eslint --ext .jsx,.js src/",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "Apache-2.0",
+  "devDependencies": {
+    "eslint": "^8.45.0",
+    "html-webpack-plugin": "^5.5.3",
+    "webpack": "^5.88.2",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
+  }
+}
