+++ src/index.html
@@ -52,13 +52,13 @@
                 }
             </script>
             
-            <script type="text/javascript" src="Common/webgl-utils.js"></script>
+            <!-- <script type="text/javascript" src="Common/webgl-utils.js"></script>
             <script type="text/javascript" src="Common/initShaders.js"></script>
             <script type="text/javascript" src="Common/MV.js"></script>
             
             <script type="text/javascript" src="Player.js"></script>
             <script type="text/javascript" src="Block.js"></script>
-            <script type="text/javascript" src="game.js"></script>
+            <script type="text/javascript" src="game.js"></script> -->
             <style>
             html {
                 width: 512px;
