+++ src/utils/MV.js
@@ -9,14 +9,14 @@
 //  Helper functions
 //
 
-function _argumentsToArray( args )
+export function _argumentsToArray( args )
 {
     return [].concat.apply( [], Array.prototype.slice.apply(args) );
 }
 
 //----------------------------------------------------------------------------
 
-function radians( degrees ) {
+export function radians( degrees ) {
     return degrees * Math.PI / 180.0;
 }
 
@@ -25,7 +25,7 @@ function radians( degrees ) {
 //  Vector Constructors
 //
 
-function vec2()
+export function vec2()
 {
     var result = _argumentsToArray( arguments );
 
@@ -37,7 +37,7 @@ function vec2()
     return result.splice( 0, 2 );
 }
 
-function vec3()
+export function vec3()
 {
     var result = _argumentsToArray( arguments );
 
@@ -50,7 +50,7 @@ function vec3()
     return result.splice( 0, 3 );
 }
 
-function vec4()
+export function vec4()
 {
     var result = _argumentsToArray( arguments );
 
@@ -69,7 +69,7 @@ function vec4()
 //  Matrix Constructors
 //
 
-function mat2()
+export function mat2()
 {
     var v = _argumentsToArray( arguments );
 
@@ -97,7 +97,7 @@ function mat2()
 
 //----------------------------------------------------------------------------
 
-function mat3()
+export function mat3()
 {
     var v = _argumentsToArray( arguments );
 
@@ -127,7 +127,7 @@ function mat3()
 
 //----------------------------------------------------------------------------
 
-function mat4()
+export function mat4()
 {
     var v = _argumentsToArray( arguments );
 
@@ -162,7 +162,7 @@ function mat4()
 //  Generic Mathematical Operations for Vectors and Matrices
 //
 
-function equal( u, v )
+export function equal( u, v )
 {
     if ( u.length != v.length ) { return false; }
 
@@ -188,7 +188,7 @@ function equal( u, v )
 
 //----------------------------------------------------------------------------
 
-function add( u, v )
+export function add( u, v )
 {
     var result = [];
 
@@ -229,7 +229,7 @@ function add( u, v )
 
 //----------------------------------------------------------------------------
 
-function subtract( u, v )
+export function subtract( u, v )
 {
     var result = [];
 
@@ -272,7 +272,7 @@ function subtract( u, v )
 
 //----------------------------------------------------------------------------
 
-function mult( u, v )
+export function mult( u, v )
 {
     var result = [];
 
@@ -321,7 +321,7 @@ function mult( u, v )
 //  Basic Transformation Matrix Generators
 //
 
-function translate( x, y, z )
+export function translate( x, y, z )
 {
     if ( Array.isArray(x) && x.length == 3 ) {
         z = x[2];
@@ -339,7 +339,7 @@ function translate( x, y, z )
 
 //----------------------------------------------------------------------------
 
-function rotate( angle, axis )
+export function rotate( angle, axis )
 {
     if ( !Array.isArray(axis) ) {
         axis = [ arguments[1], arguments[2], arguments[3] ];
@@ -367,7 +367,7 @@ function rotate( angle, axis )
 
 //----------------------------------------------------------------------------
 
-function scalem( x, y, z )
+export function scalem( x, y, z )
 {
     if ( Array.isArray(x) && x.length == 3 ) {
         z = x[2];
@@ -388,7 +388,7 @@ function scalem( x, y, z )
 //  ModelView Matrix Generators
 //
 
-function lookAt( eye, at, up )
+export function lookAt( eye, at, up )
 {
     if ( !Array.isArray(eye) || eye.length != 3) {
         throw "lookAt(): first parameter [eye] must be an a vec3";
@@ -427,7 +427,7 @@ function lookAt( eye, at, up )
 //  Projection Matrix Generators
 //
 
-function ortho( left, right, bottom, top, near, far )
+export function ortho( left, right, bottom, top, near, far )
 {
     if ( left == right ) { throw "ortho(): left and right are equal"; }
     if ( bottom == top ) { throw "ortho(): bottom and top are equal"; }
@@ -450,7 +450,7 @@ function ortho( left, right, bottom, top, near, far )
 
 //----------------------------------------------------------------------------
 
-function perspective( fovy, aspect, near, far )
+export function perspective( fovy, aspect, near, far )
 {
     var f = 1.0 / Math.tan( radians(fovy) / 2 );
     var d = far - near;
@@ -471,7 +471,7 @@ function perspective( fovy, aspect, near, far )
 //  Matrix Functions
 //
 
-function transpose( m )
+export function transpose( m )
 {
     if ( !m.matrix ) {
         return "transpose(): trying to transpose a non-matrix";
@@ -495,7 +495,7 @@ function transpose( m )
 //  Vector Functions
 //
 
-function dot( u, v )
+export function dot( u, v )
 {
     if ( u.length != v.length ) {
         throw "dot(): vectors are not the same dimension";
@@ -511,7 +511,7 @@ function dot( u, v )
 
 //----------------------------------------------------------------------------
 
-function negate( u )
+export function negate( u )
 {
     var result = [];
     for ( var i = 0; i < u.length; ++i ) {
@@ -523,7 +523,7 @@ function negate( u )
 
 //----------------------------------------------------------------------------
 
-function cross( u, v )
+export function cross( u, v )
 {
     if ( !Array.isArray(u) || u.length < 3 ) {
         throw "cross(): first argument is not a vector of at least 3";
@@ -544,14 +544,14 @@ function cross( u, v )
 
 //----------------------------------------------------------------------------
 
-function length( u )
+export function length( u )
 {
     return Math.sqrt( dot(u, u) );
 }
 
 //----------------------------------------------------------------------------
 
-function normalize( u, excludeLastComponent )
+export function normalize( u, excludeLastComponent )
 {
     if ( excludeLastComponent ) {
         var last = u.pop();
@@ -576,7 +576,7 @@ function normalize( u, excludeLastComponent )
 
 //----------------------------------------------------------------------------
 
-function mix( u, v, s )
+export function mix( u, v, s )
 {
     if ( typeof s !== "number" ) {
         throw "mix: the last paramter " + s + " must be a number";
@@ -599,7 +599,7 @@ function mix( u, v, s )
 // Vector and Matrix functions
 //
 
-function scale( s, u )
+export function scale( s, u )
 {
     if ( !Array.isArray(u) ) {
         throw "scale: second parameter " + u + " is not a vector";
@@ -618,7 +618,7 @@ function scale( s, u )
 //
 //
 
-function flatten( v )
+export function flatten( v )
 {
     if ( v.matrix === true ) {
         v = transpose( v );
@@ -666,7 +666,7 @@ var sizeof = {
 
 // printing
 
-function printm(m)
+export function printm(m)
 {
     if(m.length == 2)
     for(var i=0; i<m.length; i++)
@@ -680,14 +680,14 @@ function printm(m)
 }
 // determinants
 
-function det2(m)
+export function det2(m)
 {
 
      return m[0][0]*m[1][1]-m[0][1]*m[1][0];
 
 }
 
-function det3(m)
+export function det3(m)
 {
      var d = m[0][0]*m[1][1]*m[2][2]
            + m[0][1]*m[1][2]*m[2][0]
@@ -699,7 +699,7 @@ function det3(m)
      return d;
 }
 
-function det4(m)
+export function det4(m)
 {
      var m0 = [
          vec3(m[1][1], m[1][2], m[1][3]),
@@ -726,7 +726,7 @@ function det4(m)
 
 }
 
-function det(m)
+export function det(m)
 {
      if(m.matrix != true) console.log("not a matrix");
      if(m.length == 2) return det2(m);
@@ -738,7 +738,7 @@ function det(m)
 
 // inverses
 
-function inverse2(m)
+export function inverse2(m)
 {
      var a = mat2();
      var d = det2(m);
@@ -750,7 +750,7 @@ function inverse2(m)
      return a;
 }
 
-function inverse3(m)
+export function inverse3(m)
 {
     var a = mat3();
     var d = det3(m);
@@ -806,7 +806,7 @@ function inverse3(m)
 
 }
 
-function inverse4(m)
+export function inverse4(m)
 {
     var a = mat4();
     var d = det4(m);
@@ -914,7 +914,7 @@ function inverse4(m)
 
    return a;
 }
-function inverse(m)
+export function inverse(m)
 {
    if(m.matrix != true) console.log("not a matrix");
    if(m.length == 2) return inverse2(m);
@@ -922,7 +922,7 @@ function inverse(m)
    if(m.length == 4) return inverse4(m);
 }
 
-function normalMatrix(m, flag)
+export function normalMatrix(m, flag)
 {
     var a = mat4();
     a = inverse(transpose(m));
